#41 Unify write logic
Decouple and unify write logic between
record and collection writers
